Perf: Remove more breakpoint checking in the interpreter. Move filtering earlier in the logging chain
